Prevent persistent ArchiveDB instances from seeing stale data
PEP-249 means that all DBAPI queries are prepended with 'BEGIN
TRANSACTION', and not automatically committed until the connection is
closed. As a consequence, Postgres will maintain a view of the database
as it was when the client issued the first query. In particular, the
row counts in pg_stat_user tables never change, even as new rows are
inserted by the filter workers. Wrapping SELECT queries in an explicit
transaction seems silly, as rollback is a no-op, but it's the easiest
way to ensure that transactions do not linger forever.
directory)