Allow HEALpix searches at different resolutions
If nside < 64, the pixel number is converted to the range of nside=64
pixels that cover the requested area.  If nside > 64, the query
selects the nside=64 parent pixel (using the index), and post-filters
on the requested pixel.  nside must be a power of 2 and <= 8192,
since chealpix uses bit-twiddling in 32-bit ints internally.