Store chunks of alerts in s3
Instead of storing individual cutouts in s3, store entire chunks
of alerts.  These are only moderately larger than just the cutouts
themselves.  The alert table tracks the URI of the blob in which
the alert was stored, along with the byte range of the target block,
which can be retrieved with a range request.