Restrict programid based on token
Non-ZTF users should see only programid 1.  To accomplish this, all
bulk alert query endpoints now have an optional programid parameter,
and tokens have a partnership field, set to true if the owner's
identity is in the partnership group.  Non-partnership tokens must
set programid to 1, or omit it, in which case it defaults to 1.
Values other than 1 raise 401 Unauthorized.  Partnership queries
pass the value of programid through.

As a side-effect, fixes #5.